updated ui and cleaned up some code

+++ app/src/main/java/com/weesnerdevelopment/playingwithgames/random/circlePhysics/collision/CircleCollisionPhysicsFragment.kt
@@ -13,7 +13,7 @@ class CircleCollisionPhysicsFragment : GameFragment() {
             loopType = GameVariables.looper.value
 
             setOnClickListener {
-                for (i in 0 until GameVariables.itemsPerClick.value)
+                for (i in 0 until GameVariables.itemsPerClick.value.toInt())
                     addCircle(
                         PhysicsBall(
                             Vector(nextInt(measuredWidth), nextInt(measuredHeight)),
